# Conflicts:
#	src/main/java/com/gugucon/shopping/order/service/OrderService.java
#	src/main/java/com/gugucon/shopping/pay/service/PayService.java

+++ src/main/java/com/gugucon/shopping/rate/dto/response/RateResponse.java
@@ -10,6 +10,6 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor(access = AccessLevel.PRIVATE)
 public class RateResponse {
 
-    private int rateCount;
+    private long rateCount;
     private double averageRate;
 }
